# Tracks layer creation performance improvement