# Description
This PR improves the performance of creating a tracks layer.

I also added an asv benchmark to evaluate the improvements.

Results on `main`:
```
[100.00%] ··· ========= ============ ============ ============
              --                       n_tracks               
              --------- --------------------------------------
                 size        10          100          1000    
              ========= ============ ============ ============
                 500      10.5±3ms    13.3±0.8ms   12.7±0.2ms 
                 5000     20.8±1ms    26.8±0.6ms    72.2±2ms  
                50000     236±4ms      119±4ms      207±4ms   
                500000   13.0±0.07s   2.32±0.07s   1.27±0.04s 
               5000000     failed       failed       failed   
              ========= ============ ============ ============
```

Results on this branch:
```
[100.00%] ··· ========= ============ ============ ============
              --                       n_tracks               
              --------- --------------------------------------
                 size        10          100          1000    
              ========= ============ ============ ============
                 500     7.26±0.1ms   12.3±0.8ms   12.3±0.8ms 
                 5000     15.0±4ms    23.2±0.4ms    66.8±8ms  
                50000     70.5±1ms     91.8±6ms     189±10ms  
                500000    728±20ms     933±60ms     974±60ms  
               5000000   9.04±0.1s    9.96±0.03s   10.2±0.09s 
              ========= ============ ============ ============
```

The improvements are more noticeable in larger datasets with longer tracks (fewer `n_tracks`) ranging from 1.1x to 10x speedups.

A considerable portion of the speedup was from removing unnecessary sorting because the data was already sorted. This includes removing a check from `_validate_track_data`, which was always true because it was called only once after the data was sorted.

In my opinion, the tracks layer still needs to be reworked (maybe with a graph layer), but while we're not there yet these improvements will be useful.

## Type of change
- [x] New feature (non-breaking change which adds functionality)

# How has this been tested?
- [x] added a test to check the new `_fast_points_lookup` function
- [x] all relevant tests pass with my change
- [x] compared using asv benchmark the tracks layer creation
- [x] all tracks example are working

## Final checklist:
- [x] My PR is the minimum possible work for the desired functionality
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my feature works
- [ ] If I included new strings, I have used `trans.` to make them localizable.
      For more information see our [translations guide](https://napari.org/developers/translations.html).