@JoOkuma 2022-11-17
Thanks for the suggestions, @andy-sweet 

I updated the changes according to your comments and my replies.