@JoOkuma 2022-11-21
I added the `n_tracks=1` case and reduced the minimum required length from 10 to 5 because some parameters were never evaluated.

Here are the updated results --- I'm using a different desktop.

Main:
```
========= ========== ========== ========== =========                                                                                                                                                                                                                           
--                         n_tracks
--------- ------------------------------------------
   size       1          10        100        1000
========= ========== ========== ========== =========
    5      51.8±0ms     n/a        n/a        n/a 
    50     72.0±0ms   47.3±0ms     n/a        n/a 
   500     49.7±0ms   56.0±0ms   62.8±0ms     n/a 
   5000    110±0ms    61.4±0ms   65.5±0ms   143±0ms
  50000    1.37±0s    233±0ms    140±0ms    208±0ms
  500000    failed    10.6±0s    1.80±0s    959±0ms
 5000000    failed     failed     failed    26.6±0s
========= ========== ========== ========== =========
```

This PR:
```
benchmark_tracks_layer.TracksSuite.time_create_layer                                                                                                                                                                                                           ok              
========= ========== ========== ========== =========
--                         n_tracks                               
--------- ------------------------------------------
   size       1          10        100        1000          
========= ========== ========== ========== =========
    5      13.2±0ms     n/a        n/a        n/a           
    50     37.4±0ms   19.4±0ms     n/a        n/a           
   500     19.4±0ms   32.3±0ms   32.3±0ms     n/a   
   5000    27.6±0ms   21.4±0ms   29.0±0ms   121±0ms 
  50000    66.2±0ms   57.4±0ms   71.4±0ms   166±0ms 
  500000   658±0ms    558±0ms    539±0ms    670±0ms 
 5000000   6.62±0s    6.08±0s    6.05±0s    6.44±0s 
========= ========== ========== ========== =========
```