@andy-sweet 2022-11-22
@JoOkuma : I ran the benchmarks (by adding the the `run-benchmarks` label) out of curiosity and also to make sure they didn't take too long to run with some of the larger sizes here. Ignore the reported failure - it's actually just detecting any difference, which in this case is that performance increased!